refactor(frontend): decompose NodeDetail god component into per-tab panels (Closes #787)
Split the 2606-line NodeDetail.tsx into a thin shell plus focused, memoized
tab panels under pages/nodeDetail/: OverviewTab, DriftTab, SbomTab,
ExecutionsTab, BootstrapHistoryTab, SecretsTab, ServicesTab, ProcessesTab,
ResourcesTab, plus ConnectivityPanel and ResolvedCredentialPanel. The shell
keeps the routed export, tab navigation, shared SSH/VNC modals, the node query
and the AI-recommendation state shared by Overview and Resources.

Reuses the already-extracted Sparkline, AiRecommendationPanel, IOSTabPanel and
utils helpers. Behavior is preserved verbatim: Tailwind v4 classes, process
virtualization (useVirtualizer/sortedProcesses/processTableRef), local-timezone
date formatting, focus-trap/a11y attributes, and every query key + enabled +
refetchInterval are unchanged.

Update the #613 source-contract test to read the process assertions from the
new ProcessesTab.tsx and assert the formatLocalDateTime timestamp helper.
