+++ frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "frontend",
   "private": true,
-  "version": "0.1.575",
+  "version": "0.1.576",
   "packageManager": "pnpm@10.29.3",
   "type": "module",
   "scripts": {
