+++ tests/unit/test_processes_tab_psutil_613.py
@@ -6,7 +6,9 @@ Processes tab migrated from salt-api on-demand task to psutil read API.
 from pathlib import Path
 
 FLEET_TS = Path("frontend/src/api/fleet.ts").read_text()
-NODE_DETAIL_TSX = Path("frontend/src/pages/NodeDetail.tsx").read_text()
+# #787: NodeDetail.tsx was decomposed into per-tab panels. The Processes tab
+# now lives in its own component; its source-contract assertions read from there.
+PROCESSES_TAB_TSX = Path("frontend/src/pages/nodeDetail/ProcessesTab.tsx").read_text()
 
 
 # ── fleet.ts contracts ──────────────────────────────────────────────────────
@@ -40,68 +42,72 @@ def test_fleet_ts_ProcessStatRow_has_cpu_pct():
     assert "cpu_pct" in FLEET_TS, "ProcessStatRow must include cpu_pct field"
 
 
-# ── NodeDetail.tsx — new data source ───────────────────────────────────────
+# ── ProcessesTab.tsx — new data source ─────────────────────────────────────
 
 
 def test_nodedetal_calls_fleetApi_processStats():
-    assert "fleetApi.processStats" in NODE_DETAIL_TSX, "NodeDetail.tsx must call fleetApi.processStats"
+    assert "fleetApi.processStats" in PROCESSES_TAB_TSX, "ProcessesTab.tsx must call fleetApi.processStats"
 
 
 def test_nodedetal_uses_cpu_pct():
-    assert "cpu_pct" in NODE_DETAIL_TSX, "NodeDetail.tsx must reference cpu_pct (new field name)"
+    assert "cpu_pct" in PROCESSES_TAB_TSX, "ProcessesTab.tsx must reference cpu_pct (new field name)"
 
 
 def test_nodedetal_uses_mem_rss_bytes():
-    assert "mem_rss_bytes" in NODE_DETAIL_TSX, "NodeDetail.tsx must reference mem_rss_bytes"
+    assert "mem_rss_bytes" in PROCESSES_TAB_TSX, "ProcessesTab.tsx must reference mem_rss_bytes"
 
 
 def test_nodedetal_uses_is_llm():
-    assert "is_llm" in NODE_DETAIL_TSX, "NodeDetail.tsx must reference is_llm for LLM highlighting"
+    assert "is_llm" in PROCESSES_TAB_TSX, "ProcessesTab.tsx must reference is_llm for LLM highlighting"
 
 
 def test_nodedetal_uses_ist_timezone():
-    # #796/#877: NodeDetail no longer hardcodes a timezone. Date rendering (and
-    # the timezone suffix) is delegated to the shared formatIST/formatISTDate
-    # helper, which formats in the viewer's local timezone.
-    assert "formatIST" in NODE_DETAIL_TSX, "NodeDetail.tsx must render dates via the formatIST helper"
+    # #796/#877: the Processes tab no longer hardcodes a timezone. The "as of"
+    # collection timestamp is rendered via the shared formatLocalDateTime helper,
+    # which formats in the viewer's local timezone.
+    assert "formatLocalDateTime" in PROCESSES_TAB_TSX, (
+        "ProcessesTab.tsx must render the collected_at timestamp via the formatLocalDateTime helper"
+    )
 
 
 def test_nodedetal_has_ist_suffix():
-    # The timezone suffix is now produced by the formatIST helper rather than a
-    # literal ' IST' string embedded in the page (#796/#877).
-    assert "formatIST" in NODE_DETAIL_TSX, "NodeDetail.tsx must use formatIST so timestamps carry a timezone suffix"
+    # The timezone suffix is produced by the formatLocalDateTime helper rather
+    # than a literal ' IST' string embedded in the page (#796/#877).
+    assert "formatLocalDateTime" in PROCESSES_TAB_TSX, (
+        "ProcessesTab.tsx must use formatLocalDateTime so the timestamp carries a local-timezone rendering"
+    )
 
 
-# ── NodeDetail.tsx — old data source REMOVED ──────────────────────────────
+# ── ProcessesTab.tsx — old data source REMOVED ────────────────────────────
 
 
 def test_nodedetal_no_old_processes_read_path():
     old_path = "/api/v1/nodes/${nodeId}/processes`"
-    assert old_path not in NODE_DETAIL_TSX, "NodeDetail.tsx must not contain the old salt-task read path /processes"
+    assert old_path not in PROCESSES_TAB_TSX, "ProcessesTab.tsx must not contain the old salt-task read path /processes"
 
 
 def test_nodedetal_no_processTaskId_state():
-    assert "processTaskId" not in NODE_DETAIL_TSX, (
-        "NodeDetail.tsx must not use processTaskId state (old salt-polling approach removed)"
+    assert "processTaskId" not in PROCESSES_TAB_TSX, (
+        "ProcessesTab.tsx must not use processTaskId state (old salt-polling approach removed)"
     )
 
 
-# ── NodeDetail.tsx — controls RETAINED ─────────────────────────────────────
+# ── ProcessesTab.tsx — controls RETAINED ───────────────────────────────────
 
 
 def test_nodedetal_retains_requestProcessAction():
-    assert "requestProcessAction" in NODE_DETAIL_TSX, (
+    assert "requestProcessAction" in PROCESSES_TAB_TSX, (
         "requestProcessAction function must be retained for Stop/Suspend/Resume"
     )
 
 
 def test_nodedetal_retains_process_stop():
-    assert "process_stop" in NODE_DETAIL_TSX, "process_stop action must be retained"
+    assert "process_stop" in PROCESSES_TAB_TSX, "process_stop action must be retained"
 
 
 def test_nodedetal_retains_process_suspend():
-    assert "process_suspend" in NODE_DETAIL_TSX, "process_suspend action must be retained"
+    assert "process_suspend" in PROCESSES_TAB_TSX, "process_suspend action must be retained"
 
 
 def test_nodedetal_retains_process_resume():
-    assert "process_resume" in NODE_DETAIL_TSX, "process_resume action must be retained"
+    assert "process_resume" in PROCESSES_TAB_TSX, "process_resume action must be retained"
